fix(capabilities): rename file
