+++ packages/neon_framework/lib/src/models/app_capabilities.dart
